LUT-29293: Allow configuration of token signature check
If the tokens are signed, the algorithm must be configured in
AuthServerConf and will be checked for concordance with the received
token algorithm.
If the tokens are not signed, signed tokens will be rejected by the
parser, and the algorithm must not be configured in AuthServerConf.

This mitigated the risk that signed token are expected, but the parser
would accept valid unsigned tokens.