LUT-29769: Add a method to tell if a token is expired
Add support for setting the creation time before the request to the token endpoint
is made, so that the expiration is determined pessimistically. Expiration is determined
with reference to the token emission time, which is between the moment the request is made
and the response is received.

Add a test. To make testing more convenient, the token service is switched to be spring managed
with dependencies injected.